pyproject.toml
[tool.poetry]
name = "health-status-calculator"
version = "0.1.0"
description = "Calculates repo health status by using Github API"
authors = ["PrithviRana1 <prithvirana022@gmail.com>"]
readme = "README.md"
packages = [{include = "health_status_calculator"}]

[tool.poetry.dependencies]
python = "^3.8.1"
flake8 = "^6.0.0"
certifi = "2022.12.7"
charset-normalizer = "3.0.1"
docopt = "0.6.2"
idna = "3.4"
mccabe = "0.7.0"
pycodestyle = "2.10.0"
pyflakes = "3.0.1"
pyyaml = "6.0"
requests = "2.28.2"
yarg = "0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
